+++ src/components/gui/VirtualStick.tsx
@@ -4,7 +4,7 @@ import {
   getVectorFromSpeedAndAngle,
   getVectorSpeed,
 } from "shutils/dist/speedAngleDistance2d";
-import React, { useCallback, useEffect, useRef } from "react";
+import React, { useCallback, useEffect, useMemo, useRef } from "react";
 import { animated, useSpring } from "react-spring";
 
 export function makeVirtualStick<ConcepFuncs extends BackdopConcepFuncs>(
@@ -90,12 +90,14 @@ export function makeVirtualStick<ConcepFuncs extends BackdopConcepFuncs>(
           circleOpacity: 1.0,
           outerOpacity: 0.9,
         });
+
+        setState({ players: { main: { virtualControlsPressTime: Date.now() } } });
       },
       []
     );
 
-    useEffect(() => {
-      const pointerUpEvent = (_event?: PointerEvent) => {
+    const pointerUpEvent = useCallback(
+      (_event: React.PointerEvent<HTMLDivElement>) => {
         local.isDown = false;
         const { inputVelocity } = getState().players.main;
         opacitySpringApi.start({ circleOpacity: 0.5, outerOpacity: 0 });
@@ -116,8 +118,14 @@ export function makeVirtualStick<ConcepFuncs extends BackdopConcepFuncs>(
           // set the input velocity to 0 if the virtual stick wasn't pressed
           setState({ players: { main: { inputVelocity: { x: 0, y: 0 } } } });
         }
-      };
+        setState({
+          players: { main: { virtualControlsReleaseTime: Date.now() } },
+        });
+      },
+      []
+    );
 
+    useEffect(() => {
       // leftThumbContainer.onPointerUpObservable.add((coordinates) => {});
 
       const pointerMoveEvent = (event: PointerEvent) => {
@@ -145,57 +153,53 @@ export function makeVirtualStick<ConcepFuncs extends BackdopConcepFuncs>(
           y: event.clientY,
         };
 
-        if (local.isDown) {
-          local.xAddPos =
-            coordinates.x -
-            SIZES.leftThumbContainer * 0.5 -
-            local.leftJoystickOffset;
+        if (!local.isDown) return;
+        local.xAddPos =
+          coordinates.x -
+          SIZES.leftThumbContainer * 0.5 -
+          local.leftJoystickOffset;
 
-          local.yAddPos =
-            coordinates.y -
-            SIZES.leftThumbContainer * 0.5 -
-            local.topJoystickOffset;
+        local.yAddPos =
+          coordinates.y -
+          SIZES.leftThumbContainer * 0.5 -
+          local.topJoystickOffset;
 
-          const limitedOffset = 35;
+        const limitedOffset = 35;
 
-          const { angle, speed } = getSpeedAndAngleFromVector({
-            x: local.xAddPos,
-            y: local.yAddPos,
-          });
+        const { angle, speed } = getSpeedAndAngleFromVector({
+          x: local.xAddPos,
+          y: local.yAddPos,
+        });
 
-          let editedSpeed = speed > limitedOffset ? limitedOffset : speed;
+        let editedSpeed = speed > limitedOffset ? limitedOffset : speed;
 
-          const clampedPosition = getVectorFromSpeedAndAngle(
-            editedSpeed,
-            angle
-          );
-          const normalizedPosition = getVectorFromSpeedAndAngle(
-            editedSpeed / limitedOffset,
-            angle
-          );
+        const clampedPosition = getVectorFromSpeedAndAngle(editedSpeed, angle);
+        const normalizedPosition = getVectorFromSpeedAndAngle(
+          editedSpeed / limitedOffset,
+          angle
+        );
 
-          local.floatLeft = clampedPosition.x;
-          local.floatTop = clampedPosition.y;
+        local.floatLeft = clampedPosition.x;
+        local.floatTop = clampedPosition.y;
 
-          springApi.start({
-            position: [local.floatLeft, local.floatTop],
-            immediate: true,
-          });
+        springApi.start({
+          position: [local.floatLeft, local.floatTop],
+          immediate: true,
+        });
 
-          setState({
-            players: {
-              main: { inputVelocity: normalizedPosition },
-            },
-          });
-        }
+        setState({
+          players: {
+            main: { inputVelocity: normalizedPosition },
+          },
+        });
       };
 
-      window.addEventListener("pointerup", pointerUpEvent);
+      // window.addEventListener("pointerup", pointerUpEvent);
       // window.addEventListener("pointercancel", pointerUpEvent);
       // window.addEventListener("pointerdown", pointerDownEvent);
       window.addEventListener("pointermove", pointerMoveEvent);
       return () => {
-        window.removeEventListener("pointerup", pointerUpEvent);
+        // window.removeEventListener("pointerup", pointerUpEvent);
         // window.removeEventListener("pointercancel", pointerUpEvent);
         // window.removeEventListener("pointerdown", pointerDownEvent);
         window.removeEventListener("pointermove", pointerMoveEvent);
@@ -207,17 +211,17 @@ export function makeVirtualStick<ConcepFuncs extends BackdopConcepFuncs>(
       <div
         id="virtual-stick"
         onPointerDown={pointerDownEvent}
-        style={
-          {
-            pointerEvents: "auto",
-            position: "absolute",
-            top: 0,
-            left: 0,
-            width: "100vw",
-            height: "100vh",
-            zIndex: 1000,
-          } as any
-        }
+        onPointerUp={pointerUpEvent}
+        style={{
+          pointerEvents: "auto" as const,
+          position: "absolute" as const,
+          top: 0,
+          left: 0,
+          width: "100vw",
+          height: "100vh",
+          zIndex: 100,
+          overflow: "hidden",
+        }}
       >
         <animated.div
           ref={refs.leftThumbContainer}