+++ src/concepts/players/rules.ts
@@ -8,6 +8,7 @@ import {
 } from "shutils/dist/speedAngleDistance2d";
 import {
   BackdopConcepFuncs,
+  BackdopOptionsUntyped,
   PlaceInfoByNamePlaceholder,
 } from "../typedConcepFuncs";
 import { makeGetSceneOrEngineUtils } from "../../utils/babylonjs/getSceneOrEngine";
@@ -17,9 +18,14 @@ const LEAVE_GROUND_CANT_JUMP_DELAY = 100; // ms
 
 export function makePlayerRules<
   ConcepFuncs extends BackdopConcepFuncs,
+  BackdopOptions extends BackdopOptionsUntyped,
   CharacterName extends string,
   PlaceInfoByName extends PlaceInfoByNamePlaceholder<string>
->(concepFuncs: ConcepFuncs, placeInfoByName: PlaceInfoByName) {
+>(
+  concepFuncs: ConcepFuncs,
+  BACKDOP_OPTIONS: BackdopOptions,
+  placeInfoByName: PlaceInfoByName
+) {
   const { getRefs, getState, makeRules, setState } = concepFuncs;
 
   const globalRefs = getRefs().global.main;
@@ -98,6 +104,7 @@ export function makePlayerRules<
     }),
     whenJumpKeyPressed: addItemEffect({
       onItemEffect() {
+        if (!BACKDOP_OPTIONS.hasJumping) return;
         setState({ players: { main: { jumpButtonPressTime: Date.now() } } });
       },
       flow: "input",
@@ -105,6 +112,7 @@ export function makePlayerRules<
     }),
     whenJumpKeyReleased: addItemEffect({
       onItemEffect() {
+        if (!BACKDOP_OPTIONS.hasJumping) return;
         setState({ players: { main: { jumpButtonReleaseTime: Date.now() } } });
       },
       flow: "input",
@@ -281,6 +289,44 @@ export function makePlayerRules<
       whenToRun: "subscribe",
     }),
 
+    whenVirtualControlsPressed: addItemEffect({
+      onItemEffect({ itemRefs: playerRefs, itemName: playerName }) {
+        clearTimeoutSafe(playerRefs.canShowVirtualButtonsTimeout);
+        playerRefs.canShowVirtualButtonsTimeout = setTimeout(() => {
+          const {
+            virtualControlsPressTime,
+            virtualControlsReleaseTime,
+          } = getState().players[playerName];
+          if (virtualControlsReleaseTime > virtualControlsPressTime) return;
+          setState({
+            players: { [playerName]: { canShowVirtualButtons: true } },
+          });
+        }, 200); // wait 200 milliseconds, to prevent buttons showing from small mouse clicks
+      },
+      check: { type: "players", prop: "virtualControlsPressTime" },
+      flow: "input",
+      whenToRun: "subscribe",
+    }),
+    whenVirtualControlsReleased: addItemEffect({
+      onItemEffect({ itemRefs: playerRefs, itemName: playerName }) {
+        clearTimeoutSafe(playerRefs.canHideVirtualButtonsTimeout);
+        playerRefs.canHideVirtualButtonsTimeout = setTimeout(() => {
+          const {
+            virtualControlsPressTime,
+            virtualControlsReleaseTime,
+          } = getState().players[playerName];
+          if (virtualControlsPressTime > virtualControlsReleaseTime) return;
+          setState({
+            players: { [playerName]: { canShowVirtualButtons: false } },
+          });
+        }, 5000); // wait 5 seconds
+      },
+      check: { type: "players", prop: ["virtualControlsReleaseTime", ""] },
+      flow: "input",
+      whenToRun: "subscribe",
+    }),
+
+    // Jumping
     onEachFrame: addItemEffect({
       onItemEffect({
         newValue: inputVelocity,
@@ -300,7 +346,7 @@ export function makePlayerRules<
           getCharDollStuff(playerCharacter as CharacterName) ?? {};
         const { isJumping, isOnGround } = getState().players.main;
 
-        if (!dollRefs.checkCollisions) return;
+        // if (!dollRefs.checkCollisions) return;
 
         const { scenes } = globalRefs;
         const { meshRef } = dollRefs;
@@ -355,7 +401,7 @@ export function makePlayerRules<
             /*mesh*/ meshRef,
             /*direction*/ new Vector3(0, -0.995, 0),
             /*relativeOrigin*/ new Vector3(0, -1, 0),
-            /*length*/ 0.25
+            /*length*/ 0.3 // 0.25 meant the bird in eggventure couldn't climb the ~45degree pan
           );
 
           const pick = scene.pickWithRay(
@@ -385,7 +431,8 @@ export function makePlayerRules<
           );
         }
 
-        if (!playerMovingPaused) newPositionMoveMode = "push";
+        // if (!playerMovingPaused) newPositionMoveMode = "push";
+        newPositionMoveMode = "push";
 
         setState({ players: { main: { isOnGround: newIsOnGround } } });
       },